Update wait_move_handle to wait for transition. (#8)
* Update wait_move_handle to wait for transition

When waiting for robot idle, e.g. waiting for move command to finish,
the wait handle will move wait for a transition from busy to idle, rather than
just for idle status.

Waiting for the robot to become busy (e.g. waiting for a move to start)
should work as before. That is, it will just wait for new is_busy=True
operating state, *not* an idle-to-busy transition.

* Add back dunder members